refactor: multiset test package structure
