+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/collection/multiset/ConstMultiSetInlineMethodsTest.kt
@@ -1,7 +1,6 @@
-package xyz.lbres.kotlinutils.collection.multiset.const
+package xyz.lbres.kotlinutils.collection.multiset
 
 import org.junit.Test
-import xyz.lbres.kotlinutils.collection.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
 import xyz.lbres.kotlinutils.collection.multiset.impl.ConstMultiSetImpl
 import xyz.lbres.kotlinutils.collection.multiset.testutils.* // ktlint-disable no-wildcard-imports no-unused-imports
 import xyz.lbres.kotlinutils.internal.constants.Suppressions
